requirements.txt
accelerate==0.25.0
ciso8601==2.3.1
hydra-core==1.3.2
liac_arff==2.5.0
numpy==1.26.4
omegaconf==2.3.0
pandas==2.2.2
scipy==1.13.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
